requirements.txt
# BlendTwin Trend Query Workbench
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
